+++ compiler/optimizer/CMakeLists.txt
@@ -104,6 +104,7 @@ compiler_library(optimizer
 	${CMAKE_CURRENT_LIST_DIR}/TrivialDeadBlockRemover.cpp
 	${CMAKE_CURRENT_LIST_DIR}/FEInliner.cpp
 	${CMAKE_CURRENT_LIST_DIR}/BenefitInliner.cpp
+	${CMAKE_CURRENT_LIST_DIR}/UnsafeSubexpressionRemover.cpp
 	${CMAKE_CURRENT_LIST_DIR}/abstractinterpreter/AbsValue.cpp
 	${CMAKE_CURRENT_LIST_DIR}/abstractinterpreter/AbsOpStack.cpp
 	${CMAKE_CURRENT_LIST_DIR}/abstractinterpreter/AbsOpArray.cpp
